refactor: Inline AI conversation into workflow runner, remove phase-module indirection
Replace {module, opts} phase tuples with %Destila.Workflow.Phase{} structs.
Move chat UI to a chat_phase/1 function component in ChatComponents.
Absorb all chat event handlers into WorkflowRunnerLive.
Delete AiConversationPhase LiveComponent.
