+++ lib/destila_web/components/board_components.ex
@@ -57,7 +57,6 @@ defmodule DestilaWeb.BoardComponents do
 
   @doc """
   Returns true if the session is in a state where a ClaudeSession GenServer should be running.
-  Only AiConversationPhase phases use persistent ClaudeSession GenServers.
   """
   def should_be_alive?(%{phase_status: :processing} = session) do
     ai_phase?(session)
@@ -68,7 +67,7 @@ defmodule DestilaWeb.BoardComponents do
   defp ai_phase?(session) do
     case Destila.Workflows.phases(session.workflow_type)
          |> Enum.at(session.current_phase - 1) do
-      {DestilaWeb.Phases.AiConversationPhase, _opts} -> true
+      %Destila.Workflow.Phase{} -> true
       _ -> false
     end
   end
